Checkpoint: Updated bookkeeper
