+++ bout_runners/submitter/__init__.py
@@ -1,0 +1,1 @@
+"""Package containing the submitter modules."""
