+++ bout_runners/submitter/submitter.py
@@ -1,0 +1,24 @@
+"""Contains the submitter class."""
+
+
+# FIXME: Make this a factory pattern. Local, PBS or SLURM can be
+#  selected. Hard to figure out how to test though...
+# FIXME: Copy the run_subprocess from subprocess_functions here
+class Submitter:
+    """
+    Submits a command.
+
+    FIXME: Add variables and attributes
+
+    FIXME: Add examples
+    """
+
+    def __init__(self):
+        """
+        FIXME
+
+        Parameters
+        ----------
+        FIXME
+        """
+        pass
